Fix frame count overflowing during calculation

+++ src/main.c
@@ -455,7 +455,7 @@ int main(void)
 
 			end_time = time_us_64();
 			diff = end_time-start_time;
-			fps = (frames*1000*1000)/diff;
+			fps = ((uint64_t)frames*1000*1000)/diff;
 			printf("Frames: %u\n"
 				"Time: %lu us\n"
 				"FPS: %lu\n",
